requirements.txt
black==25.9.0
click==8.3.0
colorama==0.4.6
isort==6.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
psycopg2-binary==2.9.10
pytokens==0.1.10
